Refactor of navigation for Production
