+++ src/components/production/Money.jsx
@@ -1,15 +1,9 @@
-import { useNavigate } from "react-router-dom"
+import ProductionDashboardNavigation from "./ProductionDashboardNavigation"
 
 export default function Money() {
-    const navigate = useNavigate()
     return (
     <div className="money">
-            <button
-                onClick={() => navigate('/production')}
-                className="mb-4 text-sm text-blue-600 underline"
-            >
-                ← Back to Production Dashboard
-            </button>
+    <ProductionDashboardNavigation/>
         <h2>
             💸 Manage Monetized Tools
         </h2>
