+++ src/components/production/Orgs.jsx
@@ -1,16 +1,10 @@
-import { useNavigate } from "react-router-dom";
+import ProductionDashboardNavigation from "./ProductionDashboardNavigation";
 
 export default function Orgs() {
-    const navigate = useNavigate()
     
     return (
     <div className="p-6">
-        <button
-            onClick={() => navigate('/production')}
-            className="mb-4 text-sm text-blue-600 underline"
-        >
-            ← Back to Production Dashboard
-        </button>
+        <ProductionDashboardNavigation/>
         <h2>
         🏢 Manage Organizations
         </h2>
